remove symbol convert to uint256_t

+++ plugins/chain_plugin/chain_plugin.cpp
@@ -2208,7 +2208,7 @@ read_only::get_account_results read_only::get_account( const get_account_params&
       const auto* t_id = d.find<chain::table_id_object, chain::by_code_scope_table>(boost::make_tuple( token_code, params.account_name, N(accounts) ));
       if( t_id != nullptr ) {
          const auto &idx = d.get_index<key_value_index, by_scope_primary>();
-         auto it = idx.find(boost::make_tuple( t_id->id, core_symbol.to_symbol_code() ));
+         auto it = idx.find(boost::make_tuple( t_id->id, uint64_t(core_symbol.to_symbol_code()) ));
          if( it != idx.end() && it->value.size() >= sizeof(asset) ) {
             asset bal;
             fc::datastream<const char *> ds(it->value.data(), it->value.size());
